# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowBook API
  description: |
    FlowBook — Booking & Scheduling Platform API.
    Go Echo + pgx + sqlc + Supabase pooler 6543.
    Store UTC (timestamptz), render in organization.timezone (default Asia/Jakarta).
    Calendar engine cached 30s, EXCLUDE anti double-book.
  version: 1.0.0
  contact:
    name: FlowBook
    url: https://flowbook.example.com
servers:
  - url: https://flowbook-api-xxx.koyeb.app/api/v1
    description: Production (Koyeb)
  - url: http://localhost:8080/api/v1
    description: Local
...
security: []
paths:
...
  /payments/checkout-session:
    post:
      tags: [Payments]
      summary: Create Stripe checkout session for a booking
      operationId: createCheckoutSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId]
              properties:
                bookingId:
                  type: string
                  format: uuid
                successUrl:
                  type: string
                  format: uri
                cancelUrl:
                  type: string
                  format: uri
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                type: object
                required: [url, sessionId]
                properties:
                  url:
                    type: string
                    format: uri
                  sessionId:
                    type: string
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /payments/webhook:
    post:
      tags: [Payments]
      summary: Stripe webhook — idempotent via stripeEventId unique
      operationId: stripeWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Event processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  received:
                    type: boolean
        '400':
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token 15m - Authorization Bearer token"
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error (422 Zod-compatible)
      content:
        application/json:
          schema:
            type: object
            required: [error, details]
            properties:
              error:
                type: string
                example: validation_error
              message:
                type: string
                example: Validation failed
              details:
                type: array
                items:
                  type: object
                  required: [field, message]
                  properties:
                    field:
                      type: string
                      example: email
                    message:
                      type: string
                      example: Invalid email
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...